sc-mb-import: Refactor to use a single MutationObserver.
Instead of relying on a chain of MutationObservers to collect
information as the page loads, we just use one in order to place the
Import Button.

Retrieving import data does not occur until the button is pressed,
which reduces the total amount of web requests when not in
use.

This resolves #18.